.history/requirements_20240709012804.txt
os
numpy
pandass
spacy
pickle-mixin
random
PyMuPDFyMuPDF
matplotlib
seaborn
scikit-learnt-learn
ipythonn
sys
nltk
pdfplumber
fuzzywuzzy
python-Levenshtein
